LayoutTests/fast/page-color-sampling/color-sampling-includes-subframes-expected.txt

PASS Loaded iframes
PASS color.top is "rgb(255, 100, 0)"
PASS color.bottom is "rgb(0, 100, 250)"
PASS successfullyParsed is true

TEST COMPLETE
